+++ ModuleSettings/IdentitySwitcherModuleSettingsRepository.cs
@@ -22,7 +22,7 @@
 
 #endregion
 
-namespace DNN.Modules.IdentitySwitcher.Components
+namespace DNN.Modules.IdentitySwitcher.ModuleSettings
 {
     using DotNetNuke.Entities.Modules.Settings;
